+++ example/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example
 
+## 1.0.3
+
+### Patch Changes
+
+- Add circleSweepDegrees and startAngle props to PieChart ([#307](https://github.com/FormidableLabs/victory-native-xl/pull/307))
+
+## 1.0.2
+
+### Patch Changes
+
+- Fix yLabel width calculation to better align x-scale ([#291](https://github.com/FormidableLabs/victory-native-xl/pull/291))
+
+- Replace getTextWidth with measureText ([#290](https://github.com/FormidableLabs/victory-native-xl/pull/290))
+
 ## 1.0.1
 
 ### Patch Changes